refactor(analysis): clarify analysis package structure
- split the analysis package into app, pipeline, and ui areas
- update imports and entry points to follow the new layout
- remove unused visualization icon helpers and stale launcher sketch references
