+++ src/analysis/app/__init__.py
@@ -1,0 +1,1 @@
+"""Analysis application entry points."""
