+++ src/analysis/pipeline/__init__.py
@@ -1,0 +1,1 @@
+"""Analysis data-processing pipeline components."""
