+++ src/analysis/pipeline/frame_analyzer.py
