+++ src/analysis/pipeline/pose_optimizer.py
