+++ src/analysis/pipeline/slicer.py
