+++ src/analysis/ui/plot_manager.py
