+++ src/analysis/ui/widget_raw_data_processing.py
@@ -8,7 +8,7 @@ from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.backends.backend_qtagg import NavigationToolbar2QT as NavigationToolbar
 from matplotlib.figure import Figure
 
-from src.analysis.core.plot_manager import PlotManager
+from src.analysis.ui.plot_manager import PlotManager
 from src.analysis.ui.data_selection_dialog import DataSelectionDialog
 from src.config import config_app
 from src.config.data_columns import (
