+++ src/analysis/ui/widget_results_analyzer.py
@@ -10,7 +10,7 @@ from matplotlib.backends.backend_qtagg import FigureCanvasQTAgg as FigureCanvas
 from matplotlib.backends.backend_qtagg import NavigationToolbar2QT as NavigationToolbar
 from matplotlib.figure import Figure
 
-from src.analysis.core.plot_manager import PlotManager
+from src.analysis.ui.plot_manager import PlotManager
 from src.analysis.ui.plot_popup_dialog import PlotPopupDialog
 from src.config.data_columns import (
     DISPLAY_RESULT_COLUMNS,
